third_party/WebKit/LayoutTests/http/tests/custom/xhr-response-does-not-have-registry.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
(function() {

var t = async_test('an XHR response document must not have a registry');

var req = new XMLHttpRequest();
req.onreadystatechange = t.step_func(function() {
  if (req.readyState == 4 && req.status == 200) {
    assert_throws(
        'NOT_SUPPORTED_ERR',
        function () { req.response.registerElement('x-a'); });
    t.done();
  }
});
req.open('GET', 'resources/blank.html', true);
req.responseType = 'document';
req.send();

})();
</script>
